% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Obtain forecasts from a csv file}
\usage{
get_forecast(
  path_to_file,
  base_model = "RidgeCV",
  n_hidden_features = 5L,
  lags = 25L,
  type_pi = "gaussian",
  replications = NULL,
  h = 10L
)
}
\arguments{
\item{path_to_file}{a string; path to csv file containing time series data (examples:
https://github.com/Techtonique/datasets/tree/main/time_series/univariate)}

\item{base_model}{a string; Forecasting method to use (default is "RidgeCV"); for now scikit-learn model names.}

\item{n_hidden_features}{an integer; Number of hidden features for the model (default is 5).}

\item{lags}{an integer; Number of lags to use in the model (default is 25).}

\item{type_pi}{a string; Type of prediction interval to use (default is 'gaussian').}

\item{replications}{an integer; Number of replications for certain methods (default is None).}

\item{h}{an integer; Forecast horizon (default is 10).}
}
\description{
Obtain forecasts from a csv file
}
